+++ crates/pop-contracts/src/lib.rs
@@ -16,12 +16,14 @@ pub use call::{
 };
 pub use new::{create_smart_contract, is_valid_contract_name};
 pub use templates::{Contract, ContractType};
-pub use test::{does_contracts_node_exist, test_e2e_smart_contract, test_smart_contract};
+pub use test::{test_e2e_smart_contract, test_smart_contract};
 pub use up::{
 	dry_run_gas_estimate_instantiate, dry_run_upload, instantiate_smart_contract,
 	set_up_deployment, set_up_upload, upload_smart_contract, UpOpts,
 };
 pub use utils::{
-	contracts_node::{download_contracts_node, is_chain_alive, run_contracts_node},
+	contracts_node::{
+		does_contracts_node_exist, download_contracts_node, is_chain_alive, run_contracts_node,
+	},
 	signer::parse_hex_bytes,
 };
